Implemented backend functionality for alert notifications
Together with Sebastian Holesz:
Implemented notification service that checks users' set alerts periodically, it then adds them to the database, and sends a notification to the frontend using a websocket.
The calls from frontend for fetching, removing and marking alert notifications as viewed work. Ommited 'notifications' property on user object, in the isAuth method.